Two functions working!
